+++ Kokoro/Features/WormStatusManager.cs
@@ -4,28 +4,9 @@ using System.Linq;
 
 namespace Shockah.Kokoro;
 
-public sealed class WormStatusManager
+public sealed class WormStatusManager : IStatusLogicHook
 {
 	private static ModEntry Instance => ModEntry.Instance;
-
-	internal void OnPlayerTurnStart(State state, Combat combat)
-	{
-		int worm = combat.otherShip.Get((Status)Instance.Content.WormStatus.Id!.Value);
-		if (worm <= 0)
-			return;
-
-		var partXsWithIntent = Enumerable.Range(0, combat.otherShip.parts.Count)
-			.Where(x => combat.otherShip.parts[x].intent is not null)
-			.Select(x => x + combat.otherShip.x)
-			.ToList();
-
-		foreach (var partXWithIntent in partXsWithIntent.Shuffle(state.rngActions).Take(worm))
-			combat.Queue(new AStunPart { worldX = partXWithIntent });
-
-		if (combat.otherShip.Get(Status.timeStop) <= 0)
-			return;
-		combat.otherShip.Add((Status)Instance.Content.WormStatus.Id!.Value, -1);
-	}
 
 	internal IEnumerable<Tooltip> ModifyCardTooltips(IEnumerable<Tooltip> tooltips)
 	{
@@ -35,5 +16,34 @@ public sealed class WormStatusManager
 				glossary.vals = new object[] { "<c=boldPink>1</c>" };
 			yield return tooltip;
 		}
+	}
+
+	public void OnStatusTurnTrigger(State state, Combat combat, StatusTurnTriggerTiming timing, Ship ship, Status status, int oldAmount, int newAmount)
+	{
+		if (status != (Status)Instance.Content.WormStatus.Id!.Value || timing != StatusTurnTriggerTiming.TurnStart)
+			return;
+
+		var otherShip = ship.isPlayerShip ? combat.otherShip : state.ship;
+		int wormAmount = otherShip.Get((Status)Instance.Content.WormStatus.Id!.Value);
+		if (wormAmount <= 0)
+			return;
+
+		if (!otherShip.isPlayerShip)
+		{
+			var partXsWithIntent = Enumerable.Range(0, otherShip.parts.Count)
+				.Where(x => otherShip.parts[x].intent is not null)
+				.Select(x => x + otherShip.x)
+				.ToList();
+
+			foreach (var partXWithIntent in partXsWithIntent.Shuffle(state.rngActions).Take(wormAmount))
+				combat.Queue(new AStunPart { worldX = partXWithIntent });
+		}
+
+		combat.Queue(new AStatus
+		{
+			targetPlayer = otherShip.isPlayerShip,
+			status = (Status)Instance.Content.WormStatus.Id!.Value,
+			statusAmount = -1
+		});
 	}
 }