+++ Soggins/docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 
-## Upcoming version
+## 1.2.2
+Released 6 January 2024.
 
 * Soggins.EXE now guarantees either a Smugness Control or Harnessing Smugness as one of the choices.
 * Infinite and Recycle card traits now obey botches (the card goes to the discard pile).